refactor(reef): make settings a multi-page section
`/settings` was a leaf route rendering MCP Clients inline, while the sidebar
already modelled settings as a section with its own nav list and matched it
with `end: false`. Turn the route into a layout with children so a second
settings page has somewhere to go.

MCP Clients moves verbatim to `/settings/mcp-clients` and the section index
redirects there, so the Desktop shell lands where it always did. Nothing
changes for a browser, which still has no settings page to render.
