+++ apps/reef/app/routes/settings.tsx
@@ -1,31 +1,13 @@
-import type { Route } from './+types/settings'
-import { useFetcher, useRouteLoaderData } from 'react-router'
-import type { loader as appShellLoader } from './app-shell'
+import { Outlet } from 'react-router'
 
-import { isCoralDesktopBuild } from '@/lib/coral-desktop'
-import { Settings, SettingsHydrateFallback } from '@/views/settings/settings'
+import * as styles from '@/views/settings/settings.css'
 
-export { clientAction, clientLoader, loader } from './settings-loader'
-
-export default function SettingsRoute({ loaderData }: Route.ComponentProps) {
-  if (loaderData.runtime !== 'desktop') return null
-
-  const fetcher = useFetcher()
-  const workspaces = useRouteLoaderData<typeof appShellLoader>('routes/app-shell')?.workspaces ?? []
-  const pendingClientId = fetcher.formData?.get('clientId')
-
+export default function SettingsRoute() {
   return (
-    <Settings
-      loaderData={loaderData}
-      pendingClientIds={typeof pendingClientId === 'string' ? [pendingClientId] : []}
-      workspaces={workspaces}
-      onWorkspaceChange={(clientId, workspace) => {
-        fetcher.submit({ clientId, workspace: workspace ?? '' }, { method: 'post' })
-      }}
-    />
+    <main className={styles.page}>
+      <div className={styles.container}>
+        <Outlet />
+      </div>
+    </main>
   )
-}
-
-export function HydrateFallback() {
-  return isCoralDesktopBuild() ? <SettingsHydrateFallback /> : null
 }
