+++ apps/reef/app/routing/routemap.ts
@@ -17,6 +17,7 @@ const WORKSPACE_SOURCE_PATH = '/workspaces/:workspaceId/sources/:sourceName'
 const WORKSPACE_TRACES_PATH = '/workspaces/:workspaceId/traces'
 const WORKSPACE_TRACE_PATH = '/workspaces/:workspaceId/traces/:traceId'
 const SETTINGS_PATH = '/settings'
+const SETTINGS_MCP_CLIENTS_PATH = '/settings/mcp-clients'
 
 export const routeDefinitions = {
   home: {
@@ -38,6 +39,10 @@ export const routeDefinitions = {
   settings: {
     path: SETTINGS_PATH,
     toPath: () => SETTINGS_PATH,
+  },
+  settingsMcpClients: {
+    path: SETTINGS_MCP_CLIENTS_PATH,
+    toPath: () => SETTINGS_MCP_CLIENTS_PATH,
   },
   workspaces: {
     path: WORKSPACES_PATH,
