+++ apps/reef/app/views/settings/settings.tsx
@@ -16,50 +16,42 @@ export function Settings({
   readonly workspaces: ReadonlyArray<{ name: string }>
 }) {
   return (
-    <main className={styles.page}>
-      <div className={styles.container}>
-        <section className={styles.section}>
-          <header className={styles.sectionHeader}>
-            <Typography.HeadingLarge as="h1">MCP Clients</Typography.HeadingLarge>
-            <Typography.Body variant="secondary">
-              Choose the Coral workspace each MCP client can access.{' '}
-              <Button.ExternalLink
-                href="https://withcoral.com/docs/guides/use-coral-over-mcp"
-                size="small"
-              >
-                Learn more
-              </Button.ExternalLink>
-            </Typography.Body>
-          </header>
+    <section className={styles.section}>
+      <header className={styles.sectionHeader}>
+        <Typography.HeadingLarge as="h1">MCP Clients</Typography.HeadingLarge>
+        <Typography.Body variant="secondary">
+          Choose the Coral workspace each MCP client can access.{' '}
+          <Button.ExternalLink
+            href="https://withcoral.com/docs/guides/use-coral-over-mcp"
+            size="small"
+          >
+            Learn more
+          </Button.ExternalLink>
+        </Typography.Body>
+      </header>
 
-          <Banner>
-            This page shows only global MCP configurations. Project-specific and other connections
-            will not appear here.
-          </Banner>
+      <Banner>
+        This page shows only global MCP configurations. Project-specific and other connections will
+        not appear here.
+      </Banner>
 
-          <McpClientsList
-            clients={loaderData.mcpClients.clients}
-            error={loaderData.mcpClients.error}
-            onWorkspaceChange={onWorkspaceChange}
-            pendingClientIds={pendingClientIds}
-            workspaces={workspaces}
-          />
-        </section>
-      </div>
-    </main>
+      <McpClientsList
+        clients={loaderData.mcpClients.clients}
+        error={loaderData.mcpClients.error}
+        onWorkspaceChange={onWorkspaceChange}
+        pendingClientIds={pendingClientIds}
+        workspaces={workspaces}
+      />
+    </section>
   )
 }
 
 export function SettingsHydrateFallback() {
   return (
-    <main className={styles.page}>
-      <div className={styles.container}>
-        <section className={styles.section}>
-          <Typography.BodySmall role="status" variant="tertiary">
-            Loading MCP clients…
-          </Typography.BodySmall>
-        </section>
-      </div>
-    </main>
+    <section className={styles.section}>
+      <Typography.BodySmall role="status" variant="tertiary">
+        Loading MCP clients…
+      </Typography.BodySmall>
+    </section>
   )
 }
